Corrected model connection name in lua
